+++ HDNXUdemy/Entities/ContentCourse.cs
@@ -2,7 +2,7 @@
 {
     public class ContentCourseEntities : BaseEntities
     {
-        public int IdCourse { get; set; }
+        public long IdCourse { get; set; }
         public string? Name { get; set; }
     }
 }
