+++ HDNXUdemy/Entities/Notification.cs
@@ -2,13 +2,13 @@
 {
     public class NotificationEntities : BaseEntities
     {
-        public int? IdComment { get; set; }
+        public long? IdComment { get; set; }
 
         public int? TypeNotification { get; set; }
 
-        public int IdCourse { get; set; }
+        public long IdCourse { get; set; }
 
-        public int? IdStudent { get; set; }
+        public long? IdStudent { get; set; }
 
         public string? ShortComment { get; set; }
