+++ HDNXUdemy/Entities/PurcharseCourseDetails.cs
@@ -2,15 +2,14 @@
 {
     public class PurcharseCourseDetailsEntities : BaseEntities
     {
-        public int IdCourse { get; set; }
+        public long IdCourse { get; set; }
 
-        public int IdStudent { get; set; }
+        public long IdStudent { get; set; }
 
         public decimal? PriceOfCourse { get; set; }
 
         public decimal? PriceOfDiscount { get; set; }
 
-        public int IdPurchaseOrder { get; set; }
-
+        public long IdPurchaseOrder { get; set; }
     }
 }
