+++ HDNXUdemyServices/IServices/IHomeServices.cs
@@ -4,6 +4,6 @@ namespace HDNXUdemyServices.IServices
 {
     public interface IHomeServices
     {
-        Task<HomeModel> GetDataForHome(int? idUser);
+        Task<HomeModel> GetDataForHome(long? idUser);
     }
 }
